Changed architecture from monolithic -> serverless
