+++ src/Validation/SchemaValidation.js
@@ -1,0 +1,13 @@
+const Joi = require("joi");
+
+const registerValidation = Joi.object({
+  name: Joi.string().min(3).max(40).required(),
+  email: Joi.string().email().required(),
+  password: Joi.string().min(6).required(),
+});
+const loginValidation = Joi.object({
+  email: Joi.string().email().required(),
+  password: Joi.string().min(6).required(),
+});
+
+module.exports = { registerValidation, loginValidation };
